organize ts glue files and refactor
