+++ packages/rs-drive/src/query/mod.rs
@@ -1257,8 +1257,8 @@ impl<'a> DriveQuery<'a> {
             let element = Element::deserialize(value).unwrap();
             match element {
                 Element::Item(val, _) => values.push(val),
-                | Element::SumItem(val, _) => values.push(val.to_be_bytes().to_vec()),
-                    Element::Tree(..) | Element::SumTree(..) | Element::Reference(..) => {
+                Element::SumItem(val, _) => values.push(val.to_be_bytes().to_vec()),
+                Element::Tree(..) | Element::SumTree(..) | Element::Reference(..) => {
                     return Err(Error::GroveDB(GroveError::InvalidQuery(
                         "path query should only point to items: got trees",
                     )));
