+++ packages/wasm-dpp/lib/errors/AbstractConsensusError.ts
