+++ packages/wasm-dpp/lib/errors/DPPError.ts
