+++ packages/wasm-dpp/lib/utils/extend.ts
