+++ packages/wasm-dpp/webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const webpack = require('webpack');
 
 module.exports = {
-  entry: './index.ts',
+  entry: './lib/index.ts',
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'index.js',
